style: animated project section + refractor projects file
